Oil Market and Macroeconomic Data

A dataset containing oil market variables and macroeconomic indicators for Vector Autoregression (VAR) analysis, along with a high-frequency external instrument for identification.

Usage

kaenzig_data

Format

A data frame with X rows and 7 columns:

Details

This dataset is designed for estimating structural Vector Autoregression models of the global oil market. The high-frequency instrument can be used for external instrument identification of oil supply or demand shocks.

Examples

# Load the data
data(kaenzig_data)

# View the first few rows
head(kaenzig_data)
#> # A tibble: 6 × 8
#>   date                Oil_Price World_Oil_Prod World_Oil_Inven World_IP US_IP
#>   <dttm>                  <dbl>          <dbl>           <dbl>    <dbl> <dbl>
#> 1 1974-01-01 00:00:00      307.          1092.            628.     379.  385.
#> 2 1974-02-01 00:00:00      306.          1093.            632.     378.  385.
#> 3 1974-03-01 00:00:00      305.          1094.            631.     378.  385.
#> 4 1974-04-01 00:00:00      305.          1095.            634.     378.  385.
#> 5 1974-05-01 00:00:00      304.          1096.            638.     379.  385.
#> 6 1974-06-01 00:00:00      303.          1095.            640.     378.  385.
#> # ℹ 2 more variables: US_CPI <dbl>, iv_kanzig_final <dbl>

# Summary statistics
summary(kaenzig_data)
#>       date                       Oil_Price     World_Oil_Prod World_Oil_Inven
#>  Min.   :1974-01-01 00:00:00   Min.   :192.6   Min.   :1082   Min.   :627.7  
#>  1st Qu.:1984-12-24 06:00:00   1st Qu.:271.1   1st Qu.:1098   1st Qu.:748.9  
#>  Median :1995-12-16 12:00:00   Median :305.1   Median :1106   Median :767.4  
#>  Mean   :1995-12-16 11:00:00   Mean   :307.5   Mean   :1108   Mean   :751.2  
#>  3rd Qu.:2006-12-08 18:00:00   3rd Qu.:345.8   3rd Qu.:1120   3rd Qu.:780.4  
#>  Max.   :2017-12-01 00:00:00   Max.   :412.0   Max.   :1132   Max.   :791.4  
#>                                                                              
#>     World_IP         US_IP           US_CPI      iv_kanzig_final 
#>  Min.   :366.6   Min.   :372.2   Min.   :384.6   Min.   :-9.901  
#>  1st Qu.:396.7   1st Qu.:403.3   1st Qu.:466.0   1st Qu.: 0.000  
#>  Median :428.1   Median :432.2   Median :503.9   Median : 0.000  
#>  Mean   :429.9   Mean   :430.1   Mean   :493.7   Mean   : 0.000  
#>  3rd Qu.:464.6   3rd Qu.:457.0   3rd Qu.:531.6   3rd Qu.: 0.000  
#>  Max.   :484.4   Max.   :467.6   Max.   :551.3   Max.   : 9.383  
#>                                                  NA's   :111